# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (instructor/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
  /courses/{id}/enroll:
    post:
      operationId: enrollCourse
      tags: [courses]
      summary: Enroll in a course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrolled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /courses/{courseId}/modules:
    get:
      operationId: getCourseModules
      tags: [courses]
      summary: Get modules for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
  /modules/{moduleId}/lessons:
    get:
      operationId: getModuleLessons
      tags: [courses]
      summary: Get lessons for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
...
  /bug-reports:
    get:
      operationId: getBugReports
      tags: [bugReports]
      summary: List bug reports
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: demoAppId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bug reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BugReport"
    post:
      operationId: createBugReport
      tags: [bugReports]
      summary: Submit a bug report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBugReportRequest"
      responses:
        "201":
          description: Bug report submitted and evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BugReportResult"
  /bug-reports/{id}:
    get:
      operationId: getBugReport
      tags: [bugReports]
      summary: Get bug report details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bug report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BugReport"
...
  /instructor/courses:
    get:
      operationId: getInstructorCourses
      tags: [courses]
      summary: Get courses created by instructor
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Instructor courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          type: string
          enum: [student, instructor]
          default: student
        learningPath:
          type: string
          enum: [manual-testing, automation-testing, api-testing, full-stack-qa]
      required:
        - email
        - password
        - name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, instructor, admin]
        learningPath:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        totalScore:
          type: integer
        level:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - totalScore
        - level
        - createdAt
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [foundations, manual-testing, automation-testing, api-testing, performance-testing, mobile-testing, cicd-testing]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedHours:
          type: integer
        modulesCount:
          type: integer
        enrolledCount:
          type: integer
        rating:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        instructorId:
          type: integer
        instructorName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - difficulty
        - estimatedHours
        - modulesCount
        - enrolledCount
        - instructorId
        - instructorName
        - createdAt
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            modules:
              type: array
              items:
                $ref: "#/components/schemas/Module"
            isEnrolled:
              type: boolean
          required:
            - modules
            - isEnrolled
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [foundations, manual-testing, automation-testing, api-testing, performance-testing, mobile-testing, cicd-testing]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedHours:
          type: integer
        imageUrl:
          type: string
      required:
        - title
        - description
        - category
        - difficulty
        - estimatedHours
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        enrolledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        progress:
          type: number
      required:
        - id
        - userId
        - courseId
        - enrolledAt
        - progress
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        orderIndex:
          type: integer
        lessonsCount:
          type: integer
      required:
        - id
        - courseId
        - title
        - description
        - orderIndex
        - lessonsCount
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [concept, example, practice, challenge]
        orderIndex:
          type: integer
        estimatedMinutes:
          type: integer
        isCompleted:
          type: boolean
      required:
        - id
        - moduleId
        - title
        - type
        - orderIndex
        - estimatedMinutes
        - isCompleted
...
    CreateBugReportRequest:
      type: object
      properties:
        demoAppId:
          type: integer
        exerciseId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        stepsToReproduce:
          type: string
        expectedResult:
          type: string
        actualResult:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        environment:
          type: string
          nullable: true
        screenshotUrl:
          type: string
          nullable: true
      required:
        - demoAppId
        - title
        - description
        - stepsToReproduce
        - expectedResult
        - actualResult
        - severity
    BugReport:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        demoAppId:
          type: integer
        demoAppName:
          type: string
        exerciseId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        stepsToReproduce:
          type: string
        expectedResult:
          type: string
        actualResult:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [open, in-progress, resolved, wont-fix]
        score:
          type: integer
          nullable: true
        environment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - demoAppId
        - demoAppName
        - title
        - description
        - stepsToReproduce
        - expectedResult
        - actualResult
        - severity
        - status
        - createdAt
    BugReportResult:
      type: object
      properties:
        bugReport:
          $ref: "#/components/schemas/BugReport"
        score:
          type: integer
        feedback:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - bugReport
        - score
        - feedback
        - suggestions
...